refactor: move kvm host capacity metric in infrastructure package
